+++ sv-parser-parser/src/general/compiler_directives.rs
@@ -239,7 +239,9 @@ pub(crate) fn default_text(s: Span) -> IResult<Span, DefaultText> {
 #[packrat_parser]
 pub(crate) fn text_macro_usage(s: Span) -> IResult<Span, TextMacroUsage> {
     let (s, a) = symbol("`")(s)?;
+    begin_keywords("directive");
     let (s, b) = text_macro_identifier(s)?;
+    end_keywords();
     let (s, c) = opt(paren(list_of_actual_arguments))(s)?;
     Ok((s, TextMacroUsage { nodes: (a, b, c) }))
 }